Сделал рефакторинг кода. Разнёс часть функций по разным файлам. Перенёс логику сохранения состояния чата в класс ChatState
