+++ .gitignore
@@ -25,6 +25,10 @@
 .idea/**/dbnavigator.xml
 
 
+# Gradle and Maven with auto-import
+# When using Gradle or Maven with auto-import, you should exclude module files,
+# since they will be recreated, and may cause churn.  Uncomment if using
+# auto-import.
 # .idea/artifacts
 # .idea/compiler.xml
 # .idea/jarRepositories.xml
@@ -53,5 +57,5 @@ __pycache__
 /Lib/
 ./pyvenv.cfg
 ./docker-compose.yml
-command_count.pickle
-
+command_count.json
+users.json
